Enabling templating of insertnode_request
1. cloud-init recipes do not presently have the ability to render the
insertnode_request variable which identifies the URI at which the
node must self-register upon boot. This is critical for instances
which are created "by AWS" rather than "by Launch" as in Spot and
Auto Scaling Groups.
2. In example cloud-init config file, demonstrate how SystemD can be
used to ensure that node self-registration is repeated until success.
This allows the Launch/Tortuga service to be down for short durations
while reducing the likelihood of orphaned Spot/ASG instances.